# Don't fall-back to in-order pool when `max_streams = 1` for remote fs