--- a/tests/queries/0_stateless/02233_optimize_aggregation_in_order_prefix.sql
+++ b/tests/queries/0_stateless/02233_optimize_aggregation_in_order_prefix.sql
@@ -1,3 +1,5 @@
+-- Tags: no-s3-storage
+
 drop table if exists data_02233;
 create table data_02233 (parent_key Int, child_key Int, value Int) engine=MergeTree() order by parent_key;