--- a/tests/queries/0_stateless/02343_aggregation_pipeline.sql
+++ b/tests/queries/0_stateless/02343_aggregation_pipeline.sql
@@ -1,3 +1,5 @@
+-- Tags: no-s3-storage
+
 -- produces different pipeline if enabled
 set enable_memory_bound_merging_of_aggregation_results = 0;