--- a/tests/queries/0_stateless/02554_fix_grouping_sets_predicate_push_down.sql
+++ b/tests/queries/0_stateless/02554_fix_grouping_sets_predicate_push_down.sql
@@ -1,3 +1,5 @@
+-- Tags: no-s3-storage
+
 DROP TABLE IF EXISTS test_grouping_sets_predicate;
 
 CREATE TABLE test_grouping_sets_predicate