--- a/tests/queries/0_stateless/02582_async_reading_with_small_limit.sql
+++ b/tests/queries/0_stateless/02582_async_reading_with_small_limit.sql
@@ -1,3 +1,5 @@
+-- Tags: no-s3-storage
+
 drop table if exists t;
 
 create table t(a UInt64) engine=MergeTree order by tuple();