### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Now we use default read pool for reading from external storage when `max_streams = 1`. It is beneficial when read prefetches are enabled.