@qoega 2023-11-28
Benchmark results do not look as good as expected. In first case you made 20 consequential reads as I understand before changes. And diff is 30-100ms only.