@nickitat 2023-11-28
it represents a complex case. total number of marks is low (under 20), so there are only 2-3 tasks anyway. so new prefetches don't add match. and mark ranges are sparse within each task, so to get maximum parallelism you need to split the whole set of marks to read on very small tasks. than each one can be prefetched. I did a quick experiment and exec time indeed became less than 50ms.